dla_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.3
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
celery==5.4.0
redis==5.1.1
anthropic==0.34.2
python-dotenv==1.0.1
psutil==6.0.0
prometheus-client==0.21.0
python-multipart==0.0.9
greenlet==3.1.1
